move authenticate and getUser to servicess
