+++ client/services/Authenticate/index.ts
@@ -1,15 +1,14 @@
-import ApiClient from '../services/ApiClient'
+import ApiClient from '../ApiClient'
 import { AxiosResponse } from 'axios'
-import { UserResponse } from '../components/Home/types'
-import auth from './firebase'
+import { UserResponse } from '../../components/Home/types'
+import auth from '../../lib/firebase'
 import firebase from 'firebase/app'
-
-const client = new ApiClient()
 
 export default async function authenticate(
   currentUser: firebase.User,
   onError?: () => void
 ) {
+  const client = new ApiClient()
   const idToken = await currentUser.getIdToken()
 
   const resp: AxiosResponse<UserResponse> = await client.request(
@@ -30,21 +29,4 @@ export async function signIn(onError?: () => void): Promise<firebase.User> {
   } catch (e) {
     if (onError) onError()
   }
-}
-
-export async function getUser(
-  userId: number,
-  idToken: string,
-  onError: () => void
-): Promise<UserResponse> {
-  const resp: AxiosResponse<UserResponse> = await client.request(
-    'GET',
-    `/user/${userId}/`,
-    {
-      headers: client.setAuthHeader(idToken),
-    },
-    onError
-  )
-
-  return resp.data
 }
