+++ client/services/User/index.ts
@@ -1,0 +1,21 @@
+import ApiClient from '../../services/ApiClient'
+import { AxiosResponse } from 'axios'
+import { UserResponse } from '../../components/Home/types'
+
+export default async function getUser(
+  userId: number,
+  idToken: string,
+  onError: () => void
+): Promise<UserResponse> {
+  const client = new ApiClient()
+  const resp: AxiosResponse<UserResponse> = await client.request(
+    'GET',
+    `/user/${userId}/`,
+    {
+      headers: client.setAuthHeader(idToken),
+    },
+    onError
+  )
+
+  return resp.data
+}
